parser: split into {ty, path}.rs
